requirements.txt
streamlit
langchain
langchain-community
mistralai
python-dotenv
faiss-cpu
sentence-transformers
pypdf
tqdm



